mysql: extract binlog package
