--- go/sqltypes/replication_constants.go
@@ -1,95 +1,0 @@
-package sqltypes
-const (
-	// ComTypeDecimal is MYSQL_TYPE_DECIMAL. It is deprecated.
-	ComTypeDecimal = 0
-
-	// ComTypeTiny is MYSQL_TYPE_TINY
-	ComTypeTiny = 1
-
-	// ComTypeShort is MYSQL_TYPE_SHORT
-	ComTypeShort = 2
-
-	// ComTypeLong is MYSQL_TYPE_LONG
-	ComTypeLong = 3
-
-	// ComTypeFloat is MYSQL_TYPE_FLOAT
-	ComTypeFloat = 4
-
-	// ComTypeDouble is MYSQL_TYPE_DOUBLE
-	ComTypeDouble = 5
-
-	// ComTypeNull is MYSQL_TYPE_NULL
-	ComTypeNull = 6
-
-	// ComTypeTimestamp is MYSQL_TYPE_TIMESTAMP
-	ComTypeTimestamp = 7
-
-	// ComTypeLongLong is MYSQL_TYPE_LONGLONG
-	ComTypeLongLong = 8
-
-	// ComTypeInt24 is MYSQL_TYPE_INT24
-	ComTypeInt24 = 9
-
-	// ComTypeDate is MYSQL_TYPE_DATE
-	ComTypeDate = 10
-
-	// ComTypeTime is MYSQL_TYPE_TIME
-	ComTypeTime = 11
-
-	// ComTypeDateTime is MYSQL_TYPE_DATETIME
-	ComTypeDateTime = 12
-
-	// ComTypeYear is MYSQL_TYPE_YEAR
-	ComTypeYear = 13
-
-	// ComTypeNewDate is MYSQL_TYPE_NEWDATE
-	ComTypeNewDate = 14
-
-	// ComTypeVarchar is MYSQL_TYPE_VARCHAR
-	ComTypeVarchar = 15
-
-	// ComTypeBit is MYSQL_TYPE_BIT
-	ComTypeBit = 16
-
-	// ComTypeTimestamp2 is MYSQL_TYPE_TIMESTAMP2
-	ComTypeTimestamp2 = 17
-
-	// ComTypeDateTime2 is MYSQL_TYPE_DATETIME2
-	ComTypeDateTime2 = 18
-
-	// ComTypeTime2 is MYSQL_TYPE_TIME2
-	ComTypeTime2 = 19
-
-	// ComTypeJSON is MYSQL_TYPE_JSON
-	ComTypeJSON = 245
-
-	// ComTypeNewDecimal is MYSQL_TYPE_NEWDECIMAL
-	ComTypeNewDecimal = 246
-
-	// ComTypeEnum is MYSQL_TYPE_ENUM
-	ComTypeEnum = 247
-
-	// ComTypeSet is MYSQL_TYPE_SET
-	ComTypeSet = 248
-
-	// ComTypeTinyBlob is MYSQL_TYPE_TINY_BLOB
-	ComTypeTinyBlob = 249
-
-	// ComTypeMediumBlob is MYSQL_TYPE_MEDIUM_BLOB
-	ComTypeMediumBlob = 250
-
-	// ComTypeLongBlob is MYSQL_TYPE_LONG_BLOB
-	ComTypeLongBlob = 251
-
-	// ComTypeBlob is MYSQL_TYPE_BLOB
-	ComTypeBlob = 252
-
-	// ComTypeVarString is MYSQL_TYPE_VAR_STRING
-	ComTypeVarString = 253
-
-	// ComTypeString is MYSQL_TYPE_STRING
-	ComTypeString = 254
-
-	// ComTypeGeometry is MYSQL_TYPE_GEOMETRY
-	ComTypeGeometry = 255
-)
